Numerous improvements and fixes
Documented most of the functions and classes.

Added an option to fetch a package into a specified directory.

Added a merge_into method to the Package class, which would merge it
into a directory, and refactored repo::merge_package to use that.

Extended the package class to store category, version and slot metadata.

Added tests of the portage package source module.